+++ examples/nextjs-app-dir-v2/src/app/[[...slug]]/page.tsx
@@ -1,15 +1,17 @@
-import { Content, fetchOneEntry, isEditing, isPreviewing } from '@builder.io/sdk-react/edge';
+import { Content, fetchOneEntry, isEditing, isPreviewing } from '@builder.io/sdk-react';
 
 const BUILDER_PUBLIC_API_KEY = 'f1a790f8c3204b3b8c5c1795aeac4660';
 
 interface PageProps {
-  params: {
-    slug: string[];
-  };
+  params: { slug: string[] };
   searchParams: Record<string, string>;
 }
 
 export default async function Page(props: PageProps) {
+  // NOTE: the import must be inside the Page component itself.
+  const { initializeNodeRuntime } = await import('@builder.io/sdk-react/node/init');
+  initializeNodeRuntime();
+
   const urlPath = '/' + (props.params?.slug?.join('/') || '');
 
   const content = await fetchOneEntry({