+++ packages/sdks-tests/src/tests/dynamic-content.spec.ts
@@ -1,14 +1,15 @@
 import { expect } from '@playwright/test';
-import { findTextInPage, test } from './helpers.js';
+import { excludeTestFor, findTextInPage, test } from './helpers/index.js';
 
 test.describe('Dynamic Content', () => {
-  test('Data Bindings', async ({ page, packageName }) => {
-    // Max call stack exceeded crash.
-    test.fail(packageName === 'solid' || packageName === 'solid-start');
-
+  test('Data Bindings', async ({ page }) => {
+    test.skip(
+      excludeTestFor({ angular: true }),
+      'Data bindings are not working in Angular -- loads forever so skipping for now'
+    );
     await page.goto('/data-bindings');
 
-    await expect(page.locator(`text="1234"`).first()).toBeVisible();
+    await expect(page.locator(`text="1234"`).first()).toBeVisible({ timeout: 10000 });
     await findTextInPage({
       page,
       text: 'The Hot Wheels™ Legends Tour is Back',
@@ -23,12 +24,12 @@ test.describe('Dynamic Content', () => {
     test('static value', async ({ page }) => {
       await page.goto('/link-url');
 
-      await page.locator(`a[href="/static-url"]`).waitFor();
+      await page.locator(`a[href="/static-url"]`).waitFor({ timeout: 10000 });
     });
     test('dynamic value', async ({ page }) => {
       await page.goto('/link-url');
 
-      await page.locator(`a[href="/dynamic-url"]`).waitFor();
+      await page.locator(`a[href="/dynamic-url"]`).waitFor({ timeout: 10000 });
     });
   });
 });